Tech Debt - Create reusable components for build result details and minor UI fixes
