+++ ui/src/components/summaryitemcomponents/DetailsCard.svelte
@@ -1,8 +1,8 @@
 <script>
-  import LighthouseSummary from "../summaryitemcomponents/LighthouseSummary.svelte";
-  import CodeSummary from "../summaryitemcomponents/CodeSummary.svelte";
-  import LinkSummary from "../summaryitemcomponents/LinkSummary.svelte";
-  import ArtillerySummary from "../summaryitemcomponents/ArtillerySummary.svelte";
+  import LighthouseSummary from "./LighthouseSummary.svelte";
+  import CodeSummary from "./CodeSummary.svelte";
+  import LinkSummary from "./LinkSummary.svelte";
+  import ArtillerySummary from "./ArtillerySummary.svelte";
   import { htmlHintRules, customHtmlHintRules, RuleType } from "../../utils/utils";
 
   export let build = {};
@@ -72,7 +72,7 @@
       <div></div>
     </div>
 
-    <div class="font-sans text-lg pt-2">
+    <div class="font-sans text-lg pt-3 pb-3">
       {#if htmlRules?.selectedRules}
         <p class="inline">HTML Rules Scanned: {htmlRules.selectedRules.split(/[,]+/).length}</p>
         <span type="button" class="inline cursor-pointer" on:click={handleClick} on:keydown={handleClick}>
@@ -113,10 +113,5 @@
         {/if}
       {/if}
     </div>
-
-    <div class="py-4">
-      <p>Build Version: {val.buildVersion}</p>
-    </div>
-
   </div>
 </div>
