+++ ui/src/containers/ArtilleryReport.svelte
@@ -1,9 +1,7 @@
 <script>
   import { Navigate } from "svelte-router-spa";
   import ArtilleryDetailTable from "../components/artillerycomponents/ArtilleryDetailTable.svelte";
-  import Breadcrumbs from "../components/misccomponents/Breadcrumbs.svelte";
   import LoadingFlat from "../components/misccomponents/LoadingFlat.svelte";
-  import Tabs from "../components/misccomponents/Tabs.svelte";
   import Icon from "../components/misccomponents/Icon.svelte";
   import Toastr from "../components/misccomponents/Toastr.svelte";
   import slug from "slug";
@@ -13,9 +11,8 @@
     userSession$,
   } from "../stores";
   import { CONSTS } from "../utils/utils";
-  import CardSummary from "../components/summaryitemcomponents/CardSummary.svelte";
   import UpdateArtilleryThreshold from "../components/artillerycomponents/UpdateArtilleryThreshold.svelte";
-  import ArtilleryDetailsCard from "../components/artillerycomponents/ArtilleryDetailsCard.svelte";
+  import BuildDetailsSlot from "../components/detailslotcomponents/BuildDetailsSlot.svelte";
 
   export let currentRoute;
 
@@ -84,60 +81,53 @@
       {#await promise}
         <LoadingFlat />
       {:then data}
-        <Breadcrumbs
-          displayMode="Artillery Load Test" />
-        <br />
-
-        <CardSummary value={data.summary} />
-
-        <ArtilleryDetailsCard
-          build={data ? data : {}}
-          on:artilleryThreshold={() => showArtilleryThreshold(data.summary, $userSession$)} />
-
-        <Tabs build={data ? data : {}} displayMode="artillery" />
-
-        {#if data.summary.latencyP95 !== undefined}
-          <div class="my-4">
-            <div class="float-right">
-              <button
-                on:click={download}
-                title="Download JSON"
-                class="bg-gray-300 hover:bg-gray-400 textdark font-bold py-1 px-1
-              rounded-lg inline-flex items-center">
-                <Icon cssClass="">
-                  <path
-                    d="M4 16v1a3 3 0 003 3h10a3 3 0 003-3v-1m-4-4l-4 4m0 0l-4-4m4 4V4" />
-                </Icon>
-              </button>
+        <BuildDetailsSlot
+          {data}
+          componentType="artillery"
+        >
+          {#if data.summary.latencyP95 !== undefined}
+            <div class="my-4">
+              <div class="float-right">
+                <button
+                  on:click={download}
+                  title="Download JSON"
+                  class="bg-gray-300 hover:bg-gray-400 textdark font-bold py-1 px-1
+                rounded-lg inline-flex items-center">
+                  <Icon cssClass="">
+                    <path
+                      d="M4 16v1a3 3 0 003 3h10a3 3 0 003-3v-1m-4-4l-4 4m0 0l-4-4m4 4V4" />
+                  </Icon>
+                </button>
+              </div>
             </div>
-          </div>
 
-          <div class="grid grid-rows-1">
-            <div class="h-5" />
-          </div>
+            <div class="grid grid-rows-1">
+              <div class="h-5" />
+            </div>
 
-          {#await getAtrData}
-            <LoadingFlat />
-          {:then atrFull}
-            <ArtilleryChart value={atrFull} />
-          {/await}
+            {#await getAtrData}
+              <LoadingFlat />
+            {:then atrFull}
+              <ArtilleryChart value={atrFull} />
+            {/await}
 
-          <ArtilleryDetailTable value={data} />
-        {:else}
-          <div class="mb-6 text-center text-xl py-8">
-            <Icon cssClass="text-yellow-800 inline-block">
-              <path
-                d="M5 3v4M3 5h4M6 17v4m-2-2h4m5-16l2.286 6.857L21 12l-5.714 2.143L13
-              21l-2.286-6.857L5 12l5.714-2.143L13 3z" />
-            </Icon>
-            Artillery Load Test was not executed
-            <Icon cssClass="text-yellow-800 inline-block">
-              <path
-                d="M5 3v4M3 5h4M6 17v4m-2-2h4m5-16l2.286 6.857L21 12l-5.714 2.143L13
-              21l-2.286-6.857L5 12l5.714-2.143L13 3z" />
-            </Icon>
-          </div>
-        {/if}
+            <ArtilleryDetailTable value={data} />
+          {:else}
+            <div class="mb-6 text-center text-xl py-8">
+              <Icon cssClass="text-yellow-800 inline-block">
+                <path
+                  d="M5 3v4M3 5h4M6 17v4m-2-2h4m5-16l2.286 6.857L21 12l-5.714 2.143L13
+                21l-2.286-6.857L5 12l5.714-2.143L13 3z" />
+              </Icon>
+              Artillery Load Test was not executed
+              <Icon cssClass="text-yellow-800 inline-block">
+                <path
+                  d="M5 3v4M3 5h4M6 17v4m-2-2h4m5-16l2.286 6.857L21 12l-5.714 2.143L13
+                21l-2.286-6.857L5 12l5.714-2.143L13 3z" />
+              </Icon>
+            </div>
+          {/if}
+        </BuildDetailsSlot>
       {:catch error}
         <p class="text-red-600 mx-auto text-2xl py-8">{error.message}</p>
       {/await}
