+++ ui/src/containers/HtmlHints.svelte
@@ -3,21 +3,15 @@
     getBuildDetails,
     userSession$,
   } from "../stores";
-  import Tabs from "../components/misccomponents/Tabs.svelte";
   import { pipe, map, flatten } from "ramda";
   import HtmlErrorsTable from "../components/htmlhintcomponents/HtmlErrorsTable.svelte";
-  import Breadcrumbs from "../components/misccomponents/Breadcrumbs.svelte";
   import Toastr from "../components/misccomponents/Toastr.svelte";
   import { CONSTS, HTMLERRORS } from "../utils/utils.js";
   import { mkConfig, generateCsv, download } from "export-to-csv";
   import { Navigate } from "svelte-router-spa";
   import LoadingFlat from "../components/misccomponents/LoadingFlat.svelte";
-  import UpdateIgnoreUrl from "../components/misccomponents/UpdateIgnoreUrl.svelte";
-  import CardSummary from "../components/summaryitemcomponents/CardSummary.svelte";
-  import UpdateHtmlRules from "../components/htmlhintcomponents/UpdateHTMLRules.svelte"
-  import HtmlHintDetailsCard from "../components/htmlhintcomponents/HTMLHintDetailsCard.svelte"
-  import slug from "slug";
   import { onMount } from "svelte";
+  import BuildDetailsSlot from "../components/detailslotcomponents/BuildDetailsSlot.svelte";
 
   export let currentRoute;
 
@@ -42,13 +36,6 @@
   }
 
   let userNotLoginToast;
-  let ignoreUrlShown;
-  let urlToIgnore;
-  let scanUrl;
-  let lastBuild;
-  let threshold = {};
-  let htmlHintRulesShown;
-  let loadingHtmlHintSettings;
 
   const onDownload = data => {
     const csvConfig = mkConfig({
@@ -78,29 +65,6 @@
     download(csvConfig)(csv);
   };
 
-  const showHtmlHintThreshold = async (summary, user) => {
-    if (!user) {
-      userNotLoginToast = true;
-      return;
-    }
-    scanUrl = summary.url;
-    lastBuild = summary;
-    htmlHintRulesShown = true;
-    loadingHtmlHintSettings = true;
-    try {
-      const res = await fetch(
-        `${CONSTS.API}/api/config/${user.apiKey}/htmlhintrules/${slug(scanUrl)}`
-      );
-      const result = await res.json();
-      threshold = result || {};
-    } catch (error) {
-      console.error("error getting threshold", error);
-      threshold = {};
-    } finally {
-      loadingHtmlHintSettings = false;
-    }
-  };
-
   let htmlRules;
   const getSelectedHtmlRules = async () => {
     await promise.then((data) => {
@@ -123,23 +87,18 @@
       <LoadingFlat />
     {:then data}
 
-    <Breadcrumbs
-    displayMode="Code Issues" />
-    <br>
-    
-      <CardSummary value={data.summary} isHtmlHintComp={true}
-      on:htmlHintThreshold={() => showHtmlHintThreshold(data.summary, $userSession$)}
-      />
-    
-      <HtmlHintDetailsCard {htmlRules} build={data ? data : {}} />
-
-      <Tabs build={data ? data : {}} displayMode="code" />
-
+    <BuildDetailsSlot 
+      {data}
+      {htmlRules}
+      user={$userSession$}
+      componentType="code"
+    >
       <HtmlErrorsTable
         on:download={() => onDownload(data)}
         errors={data.htmlHint}
         codeIssues={data.codeIssues}
-        {currentRoute} />
+        {currentRoute} />  
+    </BuildDetailsSlot>
     {:catch error}
       <p class="text-red-600 mx-auto text-2xl py-8">{error.message}</p>
     {/await}
@@ -156,31 +115,5 @@
     </span>
   </p>
 </Toastr>
-
-{#if !(Object.keys(threshold).length === 0)}
-<UpdateHtmlRules url={scanUrl}
-  loading={loadingHtmlHintSettings}
-  bind:show={htmlHintRulesShown}
-  user={$userSession$}
-  {htmlRules}
-  {threshold}
-  on:updateHtmlRules={() => getSelectedHtmlRules()}
-  />
-{:else}
-<UpdateHtmlRules url={scanUrl}
-  loading={loadingHtmlHintSettings}
-  bind:show={htmlHintRulesShown}
-  user={$userSession$}
-  htmlRules={null}
-  threshold={null}
-  on:updateHtmlRules={() => getSelectedHtmlRules()}
-  />
-{/if}
-
-<UpdateIgnoreUrl
-  url={urlToIgnore}
-  {scanUrl}
-  bind:show={ignoreUrlShown}
-  user={$userSession$} />
