+++ ui/src/containers/LighthouseReport.svelte
@@ -1,60 +1,26 @@
 <script>
   import { onMount } from "svelte";
   import { Navigate } from "svelte-router-spa";
-  import Breadcrumbs from "../components/misccomponents/Breadcrumbs.svelte";
   import LoadingFlat from "../components/misccomponents/LoadingFlat.svelte";
   import Icon from "../components/misccomponents/Icon.svelte";
-  import Tabs from "../components/misccomponents/Tabs.svelte";
   import Toastr from "../components/misccomponents/Toastr.svelte";
-  import slug from "slug";
   import {
     getBuildDetails,
     userSession$,
   } from "../stores";
-  import LighthouseDetailsCard from "../components/lighthousecomponents/LighthouseDetailsCard.svelte";
-  import UpdatePerfThreshold from "../components/lighthousecomponents/UpdatePerfThreshold.svelte";
   import { CONSTS } from "../utils/utils";
-  import CardSummary from "../components/summaryitemcomponents/CardSummary.svelte";
   import { renderReport } from 'lighthouse-viewer';
+  import BuildDetailsSlot from "../components/detailslotcomponents/BuildDetailsSlot.svelte";
 
   export let currentRoute;
 
   let loading;
-
   let promise = getBuildDetails(currentRoute.namedParams.id);
   let runId;
   let userNotLoginToast;
-  let perfThresholdShown;
-  let scanUrl;
-  let loadingPerfSettings;
-  let lastBuild;
-  let threshold;
 
   const download = () => {
     window.location.href = `${CONSTS.BlobURL}/lhr/${currentRoute.namedParams.id}.json`;
-  };
-
-  const showPerfThreshold = async (summary, user) => {
-    if (!user) {
-      userNotLoginToast = true;
-      return;
-    }
-    scanUrl = summary.url;
-    lastBuild = summary;
-    perfThresholdShown = true;
-    loadingPerfSettings = true;
-    try {
-      const res = await fetch(
-        `${CONSTS.API}/api/config/${user.apiKey}/perfthreshold/${slug(scanUrl)}`
-      );
-      const result = await res.json();
-      threshold = result || {};
-    } catch (error) {
-      console.error("error getting threshold", error);
-      threshold = {};
-    } finally {
-      loadingPerfSettings = false;
-    }
   };
 
   onMount(() => {
@@ -85,19 +51,12 @@
       {#await promise}
         <LoadingFlat />
       {:then data}
-        <Breadcrumbs
-          displayMode="Lighthouse Audit" />
-        <br />
-
-        <CardSummary 
-          value={data.summary}
-          isLighthouseAudit={true}
-          on:perfThreshold={() => showPerfThreshold(data.summary, $userSession$)}
-        />
-
-        <LighthouseDetailsCard build={data ? data : {}} />
-
-        <Tabs build={data ? data : {}} displayMode="lighthouse" />
+      <BuildDetailsSlot
+        {data}
+        user={$userSession$}
+        componentType="lighthouse"
+      >
+      </BuildDetailsSlot>
       {:catch error}
         <p class="text-red-600 mx-auto text-2xl py-8">{error.message}</p>
       {/await}
@@ -120,14 +79,6 @@
     <main id="report" />
   </div>
 </div>
-
-<UpdatePerfThreshold
-  url={scanUrl}
-  loading={loadingPerfSettings}
-  {lastBuild}
-  {threshold}
-  bind:show={perfThresholdShown}
-  user={$userSession$} />
 
 <Toastr bind:show={userNotLoginToast} timeout={10000} mode="warn">
   <p>Sign in to unlock this feature!</p>
