--- ui/src/utils/Untitled-1.js
@@ -1,38 +1,0 @@
-[
-	{
-		error: 'inline-style-disabled',
-		pages: [
-			{
-				error: 'inline-style-disabled',
-				url: 'https://azuregems.io/',
-				locations: [
-					'89:35',
-					'125:41',
-					'147:41',
-					'169:41',
-					'191:41',
-					'213:41',
-					'235:41',
-					'256:41',
-					'277:41',
-					'298:41',
-				],
-			},
-			{
-				error: 'inline-style-disabled',
-				url: 'https://azuregems.io/net-5/',
-				locations: ['157:44', '222:41', '243:41'],
-			},
-		],
-	},
-	{
-		error: 'id-unique',
-		pages: [
-			{
-				error: 'id-unique',
-				url: 'https://azuregems.io/net-5/',
-				locations: ['208:21'],
-			},
-		],
-	},
-];
